vistascan-be/pyproject.toml
[project]
name = "vistascan-be"
version = "0.1.0"
description = "API for Remote Consultation platform, connecting patients with radiology experts for medical imaging studies assessments."
authors = [
    {name = "tudormihaita",email = "tudor.mihaita@ymail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "minio (>=7.2.15,<8.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "mongoengine (>=0.29.1,<0.30.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
name = "vistascan-be"
packages = [
    { include = "domain", from = "src" },
    { include = "application", from = "src" },
    { include = "infrastructure", from = "src" },
    { include = "api", from = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
